+++ pkg/webui/console/containers/gateway-onboarding-form/qr-scan-section/index.js
@@ -1,4 +1,4 @@
-// Copyright © 2022 The Things Network Foundation, The Things Industries B.V.
+// Copyright © 2024 The Things Network Foundation, The Things Industries B.V.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -13,12 +13,12 @@
 // limitations under the License.
 
 import React, { useCallback, useState } from 'react'
-import { useDispatch, useSelector } from 'react-redux'
+import { useDispatch } from 'react-redux'
+import { defineMessages } from 'react-intl'
 
-import Icon, { IconCheck, IconX } from '@ttn-lw/components/icon'
 import QRModalButton from '@ttn-lw/components/qr-modal-button'
 import { useFormContext } from '@ttn-lw/components/form'
-import Link from '@ttn-lw/components/link'
+import Icon from '@ttn-lw/components/icon'
 import ModalButton from '@ttn-lw/components/button/modal-button'
 import ButtonGroup from '@ttn-lw/components/button/group'
 
@@ -27,23 +27,27 @@ import Message from '@ttn-lw/lib/components/message'
 import attachPromise from '@ttn-lw/lib/store/actions/attach-promise'
 import sharedMessages from '@ttn-lw/lib/shared-messages'
 
-import { parseEndDeviceQRCode } from '@console/store/actions/qr-code-generator'
+// TODO: Get the correct sdk (gateway qr code generator)
+import { parseGatewayQRCode } from '@console/store/actions/qr-code-generator'
 
-import { selectDeviceBrands } from '@console/store/selectors/device-repository'
-
-import m from '../messages'
+const m = defineMessages({
+  hasGatewayQR:
+    'Does your gateway have a LoRaWAN® Gateway Identification QR Code? Scan it to speed up onboarding.',
+  gatewayGuide: 'Gateway registration help',
+  invalidQRCode:
+    'Invalid QR code data. Please note that only TTIGPRO1 Gateway Identification QR Code can be scanned. Some gateways have unrelated QR codes printed on them that cannot be used.',
+})
 
 const qrDataInitialState = {
   valid: false,
   approved: false,
   data: undefined,
-  device: undefined,
+  gateway: undefined,
 }
 
-const DeviceQRScanFormSection = () => {
+const GatewayQRScanSection = () => {
   const dispatch = useDispatch()
   const { resetForm, setValues } = useFormContext()
-  const brands = useSelector(selectDeviceBrands)
   const [qrData, setQrData] = useState(qrDataInitialState)
 
   const handleReset = useCallback(() => {
@@ -51,47 +55,24 @@ const DeviceQRScanFormSection = () => {
     setQrData(qrDataInitialState)
   }, [resetForm])
 
-  const getBrand = useCallback(
-    vendorId => {
-      const brand = brands.find(brand => brand?.lora_alliance_vendor_id === vendorId)
-
-      return brand
-    },
-    [brands],
-  )
-
   const handleQRCodeApprove = useCallback(() => {
-    const { device } = qrData
-    const { end_device } = device.end_device_template
-    const { lora_alliance_profile_ids } = end_device
-
-    const brand = getBrand(lora_alliance_profile_ids.vendor_id)
+    const { gateway } = qrData
 
     setValues(values => ({
       ...values,
       _withQRdata: true,
       ids: {
         ...values.ids,
-        join_eui: end_device.ids.join_eui,
-        dev_eui: end_device.ids.dev_eui,
-        device_id: `eui-${end_device.ids.dev_eui.toLowerCase()}`,
+        eui: gateway.gateway_eui,
       },
-      target_device_id: `eui-${end_device.ids.dev_eui.toLowerCase()}`,
       authenticated_identifiers: {
-        dev_eui: end_device.ids.dev_eui,
-        authentication_code: end_device.claim_authentication_code.value
-          ? end_device.claim_authentication_code.value
-          : '',
-        join_eui: end_device.ids.join_eui,
-      },
-      version_ids: {
-        ...values.version_ids,
-        brand_id: brand ? brand.brand_id : values.version_ids.brand_id,
+        gateway_eui: gateway.gateway_eui,
+        authentication_code: gateway.owner_token ? gateway.owner_token : '',
       },
     }))
 
     setQrData({ ...qrData, approved: true })
-  }, [getBrand, qrData, setValues])
+  }, [qrData, setValues])
 
   const handleQRCodeCancel = useCallback(() => {
     setQrData(qrDataInitialState)
@@ -100,36 +81,25 @@ const DeviceQRScanFormSection = () => {
   const handleQRCodeRead = useCallback(
     async qrCode => {
       try {
-        // Get end device template from QR code
-        const device = await dispatch(attachPromise(parseEndDeviceQRCode(qrCode)))
+        // Get gateway from QR code
+        const gateway = await dispatch(attachPromise(parseGatewayQRCode(qrCode)))
 
-        const { end_device } = device.end_device_template
-        const { lora_alliance_profile_ids } = end_device
-
-        const brand = getBrand(lora_alliance_profile_ids.vendor_id)
         const sheetData = [
           {
             header: sharedMessages.qrCodeData,
             items: [
               {
                 key: sharedMessages.claimAuthCode,
-                value: end_device.claim_authentication_code.value,
+                value: gateway.owner_token,
                 type: 'code',
                 sensitive: true,
               },
               {
-                key: sharedMessages.joinEUI,
-                value: end_device.ids.join_eui,
+                key: sharedMessages.gatewayEUI,
+                value: gateway.gateway_eui,
                 type: 'byte',
                 sensitive: false,
               },
-              {
-                key: sharedMessages.devEUI,
-                value: end_device.ids.dev_eui,
-                type: 'byte',
-                sensitive: false,
-              },
-              { key: sharedMessages.brand, value: brand?.name },
             ],
           },
         ]
@@ -137,32 +107,32 @@ const DeviceQRScanFormSection = () => {
           ...qrData,
           valid: true,
           data: sheetData,
-          device,
+          gateway,
         })
       } catch (error) {
         setQrData({ ...qrData, data: [], valid: false })
       }
     },
-    [dispatch, getBrand, qrData],
+    [dispatch, qrData],
   )
 
   return (
     <>
       {qrData.approved ? (
         <div className="mb-cs-xs">
-          <Icon icon={IconCheck} textPaddedRight className="c-bg-success-normal" />
+          <Icon icon="check" textPaddedRight className="c-success" />
           <Message content={sharedMessages.scanSuccess} />
         </div>
       ) : (
         <div className="mb-cs-xs">
-          <Message content={m.hasEndDeviceQR} />
+          <Message content={m.hasGatewayQR} />
         </div>
       )}
       <ButtonGroup>
         {qrData.approved ? (
           <ModalButton
             type="button"
-            icon={IconX}
+            icon="close"
             onApprove={handleReset}
             message={sharedMessages.qrCodeDataReset}
             modalData={{
@@ -171,27 +141,24 @@ const DeviceQRScanFormSection = () => {
               buttonMessage: sharedMessages.qrCodeDataReset,
               children: <Message content={sharedMessages.resetConfirm} component="span" />,
               approveButtonProps: {
-                icon: IconX,
+                icon: 'close',
               },
             }}
           />
         ) : (
           <QRModalButton
-            message={sharedMessages.scanEndDevice}
-            invalidMessage={m.invalidData}
+            message={sharedMessages.scanGatewayQR}
+            invalidMessage={m.invalidQRCode}
             onApprove={handleQRCodeApprove}
             onCancel={handleQRCodeCancel}
             onRead={handleQRCodeRead}
             qrData={qrData}
           />
         )}
-        <Link.DocLink className="ml-cs-xs" path="/devices/adding-devices" secondary>
-          <Message content={m.deviceGuide} />
-        </Link.DocLink>
       </ButtonGroup>
-      <hr className="mt-cs-m mb-0" />
+      <hr className="mt-cs-l mb-cs-xl" />
     </>
   )
 }
 
-export default DeviceQRScanFormSection
+export default GatewayQRScanSection